Don't throw in ppx_deriver on non-record types
Didn't work correctly with ppx_import, which is hugely usefule for this
purpose as it helps keep test code separate from production code.

There should be a way to do it still by detecting the compilation cycle,
but I was unable to get that to compile, and the error is just a minor
thing; whoever is using it should understand the purpose. Now it just
doesn't generate any code if not used on a record.

Solution: https://tarides.com/blog/2019-05-09-an-introduction-to-ocaml-ppx-ecosystem/#compatibility-with-ppx_import